Extract promptAsync to own src/util file
